adds callback capability when inserting katex blocks (#113)
We want to be able to subscribe to the event of new katex blocks being
added to the editor.

Exports a `registerInsertKatexCallback` that allows the developer to
implement a callback.

In the `Katex/Katex plugin` storybook, check if the registered callback
(`console.log`) is being called whenever a new katex block is added to
the editor.